test: update for changes to flowmodel submodule
