+++ tests/test_flowmodel/test_flowmodel_base.py
@@ -8,7 +8,8 @@ import pytest
 import torch
 from unittest.mock import create_autospec, MagicMock, patch
 
-from nessai.flowmodel import update_config, FlowModel
+from nessai.flowmodel import FlowModel
+from nessai.flowmodel import config
 
 
 @pytest.fixture()
@@ -28,39 +29,11 @@ def flow_model(flow_config, data_dim, tmpdir):
     return FlowModel(flow_config, output=output)
 
 
-def test_update_config_none():
-    """Test update config for no input"""
-    config = update_config(None)
-    assert "model_config" in config
-
-
-def test_update_config_invalid_type():
-    """Test update config when an invalid argument is specified"""
-    with pytest.raises(TypeError) as excinfo:
-        update_config(False)
-    assert "Must pass a dictionary" in str(excinfo.value)
-
-
-@pytest.mark.parametrize("noise_scale", ["auto", 4])
-def test_update_config_invalid_noise_scale(noise_scale):
-    """Assert an error is raised if noise_scale is not a float or adapative."""
-    config = {"noise_scale": noise_scale}
-    with pytest.raises(ValueError) as excinfo:
-        update_config(config)
-    assert "noise_scale must be a float or" in str(excinfo.value)
-
-
-def test_update_config_n_neurons():
-    """Assert the n_neurons is set to 2x n_inputs"""
-    config = dict(model_config=dict(n_inputs=10))
-    config = update_config(config)
-    assert config["model_config"]["n_neurons"] == 20
-
-
 def test_init_no_config(tmpdir):
     """Test the init method with no config specified"""
     output = str(tmpdir.mkdir("no_config"))
-    default_config = update_config(None)
+    default_config = config.DEFAULT_FLOW_CONFIG.copy()
+    default_config["model_config"] = config.DEFAULT_MODEL_CONFIG.copy()
     fm = FlowModel(output=output)
 
     assert fm.model_config == default_config["model_config"]
@@ -85,7 +58,7 @@ def test_initialise(model):
     model.optimiser = "adam"
     model.optimiser_kwargs = {"weights": 0.1}
     with patch(
-        "nessai.flowmodel.configure_model", return_value=("model", "cpu")
+        "nessai.flowmodel.base.configure_model", return_value=("model", "cpu")
     ) as mock:
         FlowModel.initialise(model)
     mock.assert_called_once_with(model.model_config)
@@ -206,7 +179,7 @@ def test_reset_model(model, weights, perms):
     model.optimiser_kwargs = {"lr": 0.01}
 
     with patch(
-        "nessai.flowmodel.configure_model",
+        "nessai.flowmodel.base.configure_model",
         return_value=(MagicMock, "cpu"),
     ) as mock:
         FlowModel.reset_model(model, weights=weights, permutations=perms)
